flatpak-builder: support signing build results and altering the GPG home directory
This allows signing build results with a given GPG key when exporting them to
the local repository, and allowing the user to change the GPG home directory
helps with key management in the pipeline environment. If neither option is
provided, the builds will not be signed as is the current default.